+++ layout/includes/courseindexheader.php
@@ -28,8 +28,6 @@ defined('MOODLE_INTERNAL') || die;
 
 require_once($CFG->dirroot . '/course/lib.php');
 
-user_preference_allow_ajax_update('drawer-open-index', PARAM_BOOL);
-
 if (isloggedin()) {
     $courseindexopen = (get_user_preferences('drawer-open-index', true) == true);
 } else {